+++ akshare/__init__.py
@@ -2402,9 +2402,10 @@ amac_manager_cancelled_info # 中国证券投资基金业协会-信息公示-诚
 1.9.84 fix: fix stock_hk_fhpx_detail_ths interface
 1.9.85 fix: fix option_dce_daily interface
 1.9.86 fix: fix index_kq_fz interface
+1.9.87 add: add option_minute_em interface
 """
 
-__version__ = "1.9.86"
+__version__ = "1.9.87"
 __author__ = "AKFamily"
 
 import sys
@@ -4003,6 +4004,7 @@ from akshare.option.option_finance_sina import (
     option_sse_minute_sina,
     option_sse_daily_sina,
     option_finance_minute_sina,
+    option_minute_em,
 )
 
 """