+++ docs/changelog.md
@@ -56,6 +56,10 @@
 
 ## 更新说明
 
+1.9.87 add: add option_minute_em interface
+
+    1. 新增 option_minute_em 接口
+
 1.9.86 fix: fix index_kq_fz interface
 
     1. 修复 index_kq_fz 接口
@@ -2325,6 +2329,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.9.87 add: add option_minute_em interface
 
 1.9.86 fix: fix index_kq_fz interface